양쪽 사이드 여백 색상 부여 등
1. 계정 삭제 버튼 누를 시, confirm 창 띄운 후 삭제되도록 변경
2. /utils 폴더에 서버에서 내려오는 타임 포맷 파싱하는 함수 만들기
3. 양쪽 사이드 여백 색상 부여 로직 추가와 그에 따른 Layout 구조 변경

+++ src/components/home/DeleteButton.tsx
@@ -18,7 +18,9 @@ export default function LogoutButton() {
     <button
       className='px-20 py-10 bg-red text-white rounded-8'
       onClick={() => {
-        deleteUser()
+        if (window.confirm('정말 계정을 삭제하시겠습니까?')) {
+          deleteUser()
+        }
       }}>
       회원 탈퇴
     </button>
